% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{iucn}
\alias{iucn}
\title{IUCN avaliation for frogs of Phyllomedusa genus}
\usage{
data(iucn)
}
\description{
Result of the function \code{\link{lets.iucn}} applied to Southern American frog genera Phyllomedusa in 2014.
}

