\name{lets.distmat}
\alias{lets.distmat}
\title{Geographic distance matrix}
\usage{
lets.distmat(xy)
}
\arguments{
  \item{x}{Matrix with two columns, first the longitude and
  second the latitude.}
}
\value{
Returns an object of class "dist".
}
\description{
Calculates geographic distance matrix based on a two column
matrix of x(longitude) and y(latitude).
}
\author{
Bruno Vilela & Fabricio Villalobos
}

