% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_gridirizer.R
\name{lets.gridirizer}
\alias{lets.gridirizer}
\title{Fits a grid into a PresenceAbsence object}
\usage{
lets.gridirizer(x)
}
\arguments{
\item{x}{A \code{\link{PresenceAbsence}} object.}
}
\value{
The result is a list of two objects. The first is a grid in shapefile
  format; the second is a presence-absence matrix with an additional column
  called SP_ID (shapefile cell identifier).
}
\description{
This function creates a grid in shapefile format and adds its
  cells' IDs to the presence-absence matrix. The function was created to
  facilitate the use of the PresenceAbsence object for the ones who prefer to
  work with a grid in shapefile format.
}
\examples{
\dontrun{
data(PAM)
PAM.grid <- lets.gridirizer(PAM)
names(PAM.grid)
# Grid in polygon format (can be saved in shapefile)
PAM.grid$Grid
# Presence-absence matrix (beginning only)
head(PAM.grid$Presence[, 1:5])

}

}
\seealso{
\code{\link{plot.PresenceAbsence}}

\code{\link{lets.presab.birds}}
}
\author{
Bruno Vilela
}
