% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_save_load.R
\name{lets.load}
\alias{lets.load}
\title{Load a PresenceAbsence object}
\usage{
lets.load(file)
}
\arguments{
\item{file}{a character string giving the name of the file to load.}
}
\description{
Reload PresenceAbsence objects written with the function
  \code{\link{lets.save}}.
}
\examples{
\dontrun{
data(PAM)
lets.save(PAM, file = "PAM.RData")
PAM <- lets.load(file = "PAM.RData")
}

}
\seealso{
\code{\link{lets.save}}

\code{\link{lets.presab}}
}
\author{
Bruno Vilela
}
