% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_maplizer_env.R
\name{lets.maplizer.env}
\alias{lets.maplizer.env}
\title{Map species attributes in both environmental and geographic spaces}
\usage{
lets.maplizer.env(x, y, z, func = mean, ras = TRUE, weighted = FALSE)
}
\arguments{
\item{x}{An object produced by \code{\link{lets.envpam}}.}

\item{y}{A numeric vector containing the species attributes to be summarized (e.g., description year, body size).}

\item{z}{A character vector with species names corresponding to the values in \code{y}. These names must match the species columns in the presence-absence matrices in `x`.}

\item{func}{A function used to summarize the attribute across species in each cell (e.g., \code{mean}, \code{median}, \code{sum}). Must return a single numeric value.}

\item{ras}{Logical. If \code{TRUE}, the result includes the attribute maps as \code{SpatRaster} objects.}

\item{weighted}{If TRUE, argument func is ignored, and weighted mean is
calculated. Weights are attributed to each species according to 1/N cells that the species occur.}
}
\value{
A list with the following elements:
\itemize{
 \item \code{Matrix_env}: A matrix with summarized attribute values in environmental space.
 \item \code{Matrix_geo}: A matrix with summarized attribute values in geographic space.
 \item \code{Env_Raster}: A \code{SpatRaster} with the attribute values mapped in environmental space.
 \item \code{Geo_Raster}: A \code{SpatRaster} with the attribute values mapped in geographic space.
}
}
\description{
Summarize species-level attributes (e.g., traits or conservation data) within each cell of environmental and geographic presence–absence matrices, enabling trait-based mapping across environmental gradients and geographic space.
}
\details{
This function is useful for trait-based macroecological analyses that aim to
understand how species attributes vary across environments or space. It uses
the output of \code{\link{lets.envpam}}, applies a summary function to the
trait values of all species present in each cell, and returns raster layers
for visualization.
}
\examples{
\dontrun{
# Load data
data("Phyllomedusa")
data("prec")
data("temp")
data("IUCN")

prec <- unwrap(prec)
temp <- unwrap(temp)
PAM <- lets.presab(Phyllomedusa, remove.cells = FALSE)
envs <- lets.addvar(PAM, c(temp, prec), onlyvar = TRUE)
colnames(envs) <- c("Temperature", "Precipitation")
wrld_simpl <- get(utils::data("wrld_simpl", package = "letsR"))
PAM <- lets.pamcrop(PAM, vect(wrld_simpl))

# Create environmental PAM
res <- lets.envpam(PAM, envs, remove.cells = FALSE)

# Map mean description year
res_map <- lets.maplizer.env(res, 
                             y = IUCN$Description_Year,
                             z = IUCN$Species)

# Plotting trait maps
lets.plot.envpam(res_map)
}

}
\seealso{
\code{\link{lets.envpam}}, \code{\link{lets.maplizer}}, \code{\link{lets.plot.envpam}}
}
\author{
Bruno Vilela
}
