% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_maplizer.R
\name{lets.maplizer}
\alias{lets.maplizer}
\title{Create a matrix summarizing species' attributes within 
cells of a PresenceAbsence object}
\usage{
lets.maplizer(x, y, z, func = mean, ras = FALSE, weighted = FALSE)
}
\arguments{
\item{x}{A \code{\link{PresenceAbsence}} object.}

\item{y}{Species attribute to be considered. It must be a numeric attribute.}

\item{z}{Species names in the same order as the attributes and exactly the 
same as named in the \code{PresenceAbsence} object.}

\item{func}{A function to summarize the species' attribute in each cell (the function must return only one value).}

\item{ras}{If \code{TRUE} the raster object will be returned 
together with the matrix.}

\item{weighted}{If TRUE, argument func is ignored, and weighted mean is
calculated. Weights are attributed to each species according to 1/N cells
that the species occur.}
}
\value{
The result can be both a \code{matrix} or a \code{list} containing 
the follow objects:

\strong{Matrix}: a \code{matrix} object with the cells' geographic 
coordinates and the summarized species' attributes within them.

\strong{Raster}: The summarized species' attributed mapped in a 
\code{SpatRaster} object.
}
\description{
Summarize species attributes per cell in a presence-absence matrix.
}
\examples{
\dontrun{
data(PAM)
data(IUCN)
trait <- IUCN$Description_Year
resu <- lets.maplizer(PAM, trait, PAM$S, ras = TRUE)
head(resu$Matrix)
plot(resu$Raster, xlab = "Longitude", ylab = "Latitude", 
main = "Mean description year per site")

}

}
\seealso{
\code{\link{lets.presab}}

\code{\link{lets.presab.birds}}
}
\author{
Bruno Vilela
}
