% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_save_load.R
\name{lets.save}
\alias{lets.save}
\title{Save a PresenceAbsence object}
\usage{
lets.save(pam, ...)
}
\arguments{
\item{pam}{A PresenceAbsence object.}

\item{...}{other arguments passed to the function \code{\link{save}}}
}
\description{
Save an external representation of a PresenceAbsence object to
  the specified file. The object can be read back from the file at a later
  date by using the function \code{\link{lets.load}}.
}
\examples{
\dontrun{
data(PAM)
lets.save(PAM, file = "PAM.RData")
PAM <- lets.load(file = "PAM.RData")
}

}
\seealso{
\code{\link{lets.presab}}

\code{\link{lets.presab.birds}}
}
\author{
Bruno Vilela
}
