% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_subsetPAMenv.R
\name{lets.subsetPAMenv}
\alias{lets.subsetPAMenv}
\title{Subset an environmental PAM by species names}
\usage{
lets.subsetPAMenv(x, species, remove.cells = TRUE)
}
\arguments{
\item{x}{An environmental PAM produced by \code{\link{lets.envpam}}.}

\item{species}{A character vector with species names to retain in the PAM.}

\item{remove.cells}{Logical. If \code{TRUE}, rows with zero richness after
subsetting are removed from matrices. Default is
\code{TRUE}.}
}
\value{
A list with the same structure as the input object:
\itemize{
  \item \code{x[[1]]}: the subsetted env PAM table.
  \item \code{x[[2]]}: the subsetted geo PAM table.
  \item \code{x[[3]]}: the raster associated with env PAM, updated with
  recalculated richness values.
  \item \code{x[[4]]}: the raster associated with geo PAM, updated with
  recalculated richness values.
}
}
\description{
Subsets an environmental-space presence-absence matrix (PAM) generated by
\code{\link{lets.envpam}} by retaining only a selected set of species.
Species not included in \code{species} are removed from both tabular
components of the object, and species richness is recalculated and written
back to the corresponding raster layers.
}
\details{
Optionally, cells with zero richness after subsetting can be removed from the
tabular components.


If none of the supplied species names match the species present in the PAM,
the function stops with an error before modifying the object.
}
\examples{
\dontrun{
data("Phyllomedusa")
data("prec")
data("temp")

prec <- unwrap(prec)
temp <- unwrap(temp)

PAM <- lets.presab(Phyllomedusa, remove.cells = FALSE)
envs <- lets.addvar(PAM, c(temp, prec), onlyvar = TRUE)
colnames(envs) <- c("Temperature", "Precipitation")
wrld_simpl <- get(utils::data("wrld_simpl", package = "letsR"))
PAM <- lets.pamcrop(PAM, vect(wrld_simpl))

res <- lets.envpam(PAM, envs)

lets.plot.envpam(
  x = res,
  species = NULL,
  cell_id_env = NULL,
  cell_id_geo = NULL,
  plot.grid = TRUE,
  world = TRUE,
  mar = c(4, 4, 4, 4)
)

sub <- sample(PAM$Species_name, 10)
pam_sub <- lets.subsetPAMenv(res, sub)
lets.plot.envpam(pam_sub)
}

}
\seealso{
\code{\link{lets.envpam}}, \code{\link{lets.plot.envpam}},
  \code{\link{lets.presab}}, \code{\link{lets.addvar}}
}
