\docType{package}
\name{letsR}
\alias{letsR}
\alias{letsR-package}
\title{Tools for macroecological data handling}
\description{
This package is being developed at LETS (Laboratorio de
Ecologia Teorica e Sintese) in UFG (Universidade Federal de
Goias, Brazil) as part of a larger project looking to
integrate recent advances in macroecological data handling
and processing. Its purpose is to make some of our commonly
used but unavailable R functions available to the larger
public. The package provides useful R functions for
researchers in ecology and evolution. The letsR package is
in continuous development and suggestions are more than
welcome! We hope you enjoy it and find it useful.
}
\details{
\tabular{ll}{ Package: \tab lestR\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2014-05-05\cr License: \tab
GPL-2\cr }
}
\author{
Bruno Vilela \email{brunovilelasilva@hotmal.com}

Fabricio Villalobos \email{fabricio.villalobos@gmail.com}
}
\keyword{package}

