% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{prec}
\alias{prec}
\title{Annual Precipitation raster for the world.}
\format{
A PackedSpatRaster object.
}
\source{
Data was modified from WorldClim (\url{https://worldclim.org/}, downloaded 10/2024).

Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. 
Very high resolution interpolated climate surfaces for global land areas. 
International Journal of Climatology 25: 1965-1978.
}
\usage{
prec
}
\description{
Annual Precipitation in mm for the world in 10 arc min of resolution.
}
\keyword{datasets}
