\name{print.PresenceAbsence}
\alias{print.PresenceAbsence}
\title{Print for object of class PresenceAbsence}
\usage{
\method{print}{PresenceAbsence}(x, \dots)
}
\arguments{
  \item{x}{an object of class PresenceAbsence (see function
  presab).}

  \item{...}{Other print parameters.}
}
\description{
Print for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}

