% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.summary.PresenceAbsence}
\alias{print.summary.PresenceAbsence}
\title{Print summary for object of class PresenceAbsence}
\usage{
\method{print.summary}{PresenceAbsence}(x, \dots)
}
\arguments{
\item{x}{an object of class PresenceAbsence (see function presab).}

\item{...}{Other print parameters.}
}
\description{
Print summary for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}

