\name{summary.PresenceAbsence}
\alias{summary.PresenceAbsence}
\title{Summary for object of class PresenceAbsence}
\usage{
\method{summary}{PresenceAbsence}(object, \dots)
}
\arguments{
  \item{object}{an object of class PresenceAbsence (see
  function presab).}

  \item{...}{Other summary parameters.}
}
\description{
Summary for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}

