% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{IUCN}
\alias{IUCN}
\title{IUCN avaliation for frogs of Phyllomedusa genus}
\usage{
data(IUCN)
}
\description{
Result of the function \code{\link{lets.iucn}} 
applied to Southern American frog genera Phyllomedusa in 2014.
}
