% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{IUCN}
\alias{IUCN}
\title{IUCN evaluation for frogs of the genus Phyllomedusa}
\format{
A data frame with 32 rows and 7 columns:
\describe{
 \item{Species}{Scientific name}
 \item{Family}{Family}
 \item{Status}{Red List Status}
 \item{Criteria}{Criteria for listing as threatened}
 \item{Population}{Population trend per IUCN}
 \item{Description_Year}{Year described}
 \item{Country}{Presence in country}
}
}
\source{
IUCN - \url{https://www.iucnredlist.org/}. 2014.
}
\usage{
IUCN
}
\description{
Result of the (deprecated) function \code{\link{lets.iucn}} 
applied to South American frog genus Phyllomedusa in 2014.
}
\keyword{datasets}
