% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lets.classvar}
\alias{lets.classvar}
\title{Fucntion to count for frequency distribution along a variable}
\usage{
lets.classvar(x, groups="default", pos, xy)
}
\arguments{
\item{x}{Presence-absence matrix with a unique variable added.}

\item{groups}{The number of classes wanted for the variable. Default calculates the number of classes default for a histogram (\code{\link{hist}}).}

\item{pos}{Column position of the variables.}

\item{xy}{Logical, if \code{TRUE} the matrix contains the coordinates in the first two columns.}
}
\value{
A matrix with species in rows and class in columns.
}
\description{
Based on a species Presence-Absence matrix including variables of interest (see \code{\link{lets.addvar}}), the function divides a continuous variable into classes and counts the frequency of each class in each species' range.
}
\examples{
\dontrun{
data(PAM)
data(temp)
pamvar <- lets.addvar(PAM, temp)
head(pamvar)
resu <- lets.classvar(x=pamvar, pos=ncol(pamvar), xy=TRUE)
}
}
\author{
Bruno Vilela
}
\references{
Morales-Castilla et al. 2013. Range size patterns of New World oscine passerines (Aves): insights from differences among migratory and sedentary clades. Journal of Biogeography, 40, 2261-2273.
}

