% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lets.distmat}
\alias{lets.distmat}
\title{Geographic distance matrix}
\usage{
lets.distmat(xy, count=TRUE)
}
\arguments{
\item{xy}{Matrix with two columns, first the longitude and second the latitude.}

\item{count}{Logical, if \code{TRUE} a counting window will open.}
}
\value{
Returns an object of class "dist".
}
\description{
Calculates geographic distance matrix based on a two column matrix of x(longitude) and y(latitude).
}
\author{
Bruno Vilela & Fabricio Villalobos
}

