% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_envpam.R
\name{lets.envpam}
\alias{lets.envpam}
\title{Create a presence–absence matrix in environmental space}
\usage{
lets.envpam(
  pam,
  envs,
  n_bins = 30,
  remove.cells = TRUE,
  remove.sp = TRUE,
  count = FALSE
)
}
\arguments{
\item{pam}{A `PresenceAbsence` object, typically created using
\code{\link{lets.presab}}.}

\item{envs}{A two-column matrix or data frame with continuous environmental
variables corresponding to the coordinates in the PAM. The first column will
be used as the x-axis and the second as the y-axis in the environmental
space.}

\item{n_bins}{Number of bins used to discretize each environmental variable.
Default is 30.}

\item{remove.cells}{Logical. Should cells with no species be removed from the
final matrix?}

\item{remove.sp}{Logical. Should species with no occurrences in environmental
space be removed?}

\item{count}{Logical. If TRUE, displays a progress bar for species processing.}
}
\value{
A list with the following elements:
\itemize{
 \item \code{Presence_and_Absence_Matrix_env}: A matrix of species presences across environmental bins.
 \item \code{Presence_and_Absence_Matrix_geo}: The original PAM coordinates associated with environmental cells.
 \item \code{Env_Richness_Raster}: A raster layer of species richness in environmental space.
 \item \code{Geo_Richness_Raster}: The original species richness raster in geographic space.
}
}
\description{
Transform a presence–absence matrix (PAM) based on geographic
 coordinates into a new PAM structured in environmental space. The function
 rasterizes the environmental variable space (based on two continuous
 environmental predictors), and assigns species presences to binned
 environmental conditions, producing a species richness raster in
 environmental space.
}
\details{
This function projects species occurrences into a two-dimensional
 environmental space, facilitating ecological analyses that depend on
 environmental gradients. The environmental space is discretized into a
 regular grid (determined by \code{n_bins}), and each cell is assigned the
 number of species occurring under those environmental conditions.
}
\examples{
\dontrun{
# Load data
data("Phyllomedusa")
data("prec")
data("temp")

prec <- unwrap(prec)
temp <- unwrap(temp)
PAM <- lets.presab(Phyllomedusa, remove.cells = FALSE)
envs <- lets.addvar(PAM, c(temp, prec), onlyvar = TRUE)
colnames(envs) <- c("Temperature", "Preciptation")
wrld_simpl <- get(utils::data("wrld_simpl", package = "letsR"))
PAM <- lets.pamcrop(PAM, vect(wrld_simpl))

# Run function
res <- lets.envpam(PAM, envs)

# Plot results
lets.plot.envpam(x = res,
            species = NULL,
            cell_id_env = NULL,
            cell_id_geo = NULL,
            T,
            world = TRUE,
            mar = c(4, 4, 4, 4))
}

}
\seealso{
\code{\link{lets.presab}}, \code{\link{lets.addvar}},
 \code{\link{lets.plot.envpam}}
}
\author{
Bruno Vilela
}
