% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lets.field}
\alias{lets.field}
\title{Create species' variables based on co-occurrence among species within focal ranges}
\usage{
lets.field(x, y, z, weigth=TRUE, count=FALSE)
}
\arguments{
\item{x}{A PresenceAbsence object.}

\item{y}{Species attribute to be considered. It must be a numeric attribute.}

\item{z}{Species names in the same order as the attributes.}

\item{weigth}{If \code{TRUE} the value is weighted by species' range size,
if \code{FALSE} the value is the mean of all species that co-occur with the
 focal species.}

\item{count}{Logical, if \code{TRUE} a counting window will open.}
}
\description{
Create species' attributes based on the attribute values of species co-occurring within individual ranges.
}
\details{
If the species do not co-occur with any other species NaN will be returned.
}
\examples{
\dontrun{
data(PAM)
range <- lets.rangesize(x=PAM, units="cell")
field <- lets.field(PAM, range, PAM$S, weigth=TRUE)
}
}
\author{
Bruno Vilela
}
\seealso{
\code{\link{lets.presab.birds}}

\code{\link{lets.presab}}
}

