% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lets_gridirizer.R
\name{lets.gridirizer}
\alias{lets.gridirizer}
\title{Fits a PresenceAbsence object into a shapefile grid}
\usage{
lets.gridirizer(x)
}
\arguments{
\item{x}{A PresenceAbsence object.}
}
\value{
The result is a list of two objects. The first is a grid in shapefile format;
the second is a presence absence matrix with an aditional column called SP_ID (shapefile cell identifier).
}
\description{
This function create a grid in shapefile format and add its IDs to the presence absence matrix.
The function was created to make easier the use of Presence Absence Matrix for the ones who prefer to work into a grid in shapefile.
}
\author{
Bruno Vilela
}
\seealso{
\code{\link{plot.PresenceAbsence}}

\code{\link{lets.presab.birds}}
}

