% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_midpoint.R
\name{lets.midpoint}
\alias{lets.midpoint}
\title{Compute the midpoint of species' geographic ranges}
\usage{
lets.midpoint(pam, planar = FALSE, method = "PC")
}
\arguments{
\item{pam}{A presence-absence \code{matrix} (sites in the rows and species
in the columns, with the first two columns containing the longitudinal
and latitudinal coordinates, respectively), or an object of class 
\code{\link{PresenceAbsence}}.}

\item{planar}{Logical, if \code{FALSE} the coordinates are in Longitude/Latitude.
If \code{TRUE} the coordinates are planar.}

\item{method}{Default option, "PC" (polygon centroid) will generate a polygon from the raster, 
and calculate the centroid of this polygon,
using the function centroid of the package geosphere. 
Note that "PC" will only work for PresenceAbsence objects and planar = FALSE.
Users can also choose the geographic midpoint, 
using the option "GM". "GM" will create a bouding box across the extremes of the 
distribution and calculate the centroid.
Alternatively, the midpoint can be calculated as the point
that minimize the distance between all cells of the PAM (center of minimun distance),
using the method "CMD".}
}
\value{
A \code{data.frame} containing the species' names and geographic coordinates 
(longitude [x], latitude [y]) of species' midpoints.
}
\description{
Calculate species' distributional midpoint from a 
presence-absence matrix.
}
\examples{
\dontrun{
data(PAM)
mid <- lets.midpoint(PAM)
mid2 <- lets.midpoint(PAM, method = "GM")
mid3 <- lets.midpoint(PAM, method = "CMD")
mid4 <- lets.midpoint(PAM, method = "GM", planar = TRUE)
mid5 <- lets.midpoint(PAM, method = "CMD", planar = TRUE)
for (sp in 1:nrow(mid)) {
 #sp = 4 # Or choose a line or species
 plot(PAM, name = mid[sp, 1])
 points(mid[sp, -1], col = adjustcolor("blue", .8), pch = 20, cex = 1.5)
 points(mid2[sp, -1], col = adjustcolor("green", .8), pch = 20, cex = 1.5)
 points(mid3[sp, -1], col = adjustcolor("yellow", .8), pch = 20, cex = 1.5)
 points(mid4[sp, -1], col = adjustcolor("purple", .8), pch = 20, cex = 1.5)
 points(mid5[sp, -1], col = adjustcolor("orange", .8), pch = 20, cex = 1.5)
 Sys.sleep(1)
}
} 

}
\seealso{
\code{\link{lets.presab}}

\code{\link{lets.presab.birds}}
}
\author{
Fabricio Villalobos & Bruno Vilela
}
