\name{lets.midpoint}
\alias{lets.midpoint}
\title{Species geographical range midpoint}
\usage{
lets.midpoint(pam, planar=FALSE)
}
\arguments{
  \item{pam}{A presence absence matrix, or an object of
  class PresenceAbsence.}

  \item{planar}{Logical, if FALSE the coordinates are in
  Longitude/Latitude. If TRUE the coordinates are planar.}
}
\value{
A matrix containing the point by species.
}
\description{
Get species geographical range midpoint from a
presence-absence matrix.
}
\author{
Fabricio Villalobos & Bruno Vilela
}
\seealso{
lets.presab
}

