% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lets_envcells.R
\name{lets.plot.envcells}
\alias{lets.plot.envcells}
\title{Map envPAM descriptors}
\usage{
lets.plot.envcells(
  x,
  y,
  ras = FALSE,
  plot_ras = TRUE,
  mfrow = c(4, 4),
  which.plot = NULL,
  col_func = NULL,
  ...
)
}
\arguments{
\item{x}{The envPAM list returned by \code{\link{lets.envpam}} (must include
\code{$Env_Richness_Raster} and \code{$Presence_and_Absence_Matrix_env}).}

\item{y}{A \code{data.frame} returned by \code{\link{lets.envcells}} with
one row per environmental cell (aligned with \code{x}).}

\item{ras}{Logical; if \code{TRUE}, returns a named list of \pkg{terra} SpatRaster
layers corresponding to each column plotted. Default \code{FALSE}.}

\item{plot_ras}{Logical; if \code{TRUE}, the function plot the graphs.
Default \code{TRUE}.}

\item{mfrow}{A vector of the form c(nr, nc). The figures will be drawn
in an nr-by-nc array on the device by rows as in par documentation.}

\item{which.plot}{Indicate the number of the columns in y to be ploted.}

\item{col_func}{A custom color ramp palette function to use for plotting variables (e.g., from \code{colorRampPalette}).}
}
\value{
Invisibly returns \code{NULL}. If \code{ras = TRUE}, returns a named list of
\link[terra]{SpatRaster} objects corresponding to each descriptor column.
}
\description{
Plots each column of the descriptor table returned by
\code{\link{lets.envcells}} back onto the environmental richness raster
embedded in the envPAM object. Rows with zero frequency are masked as \code{NA}.
}
\details{
Plot environmental descriptors over the environmental raster grid


Each descriptor column is assigned as values of the environmental raster template
and plotted sequentially. The plotting grid defaults to \code{par(mfrow = c(4,4))}.
}
\examples{
\dontrun{
data("Phyllomedusa"); data("prec"); data("temp")
prec <- unwrap(prec); temp <- unwrap(temp)
PAM  <- lets.presab(Phyllomedusa, remove.cells = FALSE)
envs <- lets.addvar(PAM, c(temp, prec), onlyvar = TRUE)
colnames(envs) <- c("Temperature", "Precipitation")
wrld_simpl <- get(utils::data("wrld_simpl", package = "letsR"))
PAM <- lets.pamcrop(PAM, terra::vect(wrld_simpl))
res <- lets.envpam(PAM, envs, n_bins = 30)
out <- lets.envcells(res, perc = 0.2)
lets.plot.envcells(res, out)
}

}
