% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lets.rangesize}
\alias{lets.rangesize}
\title{Calculate species range size}
\usage{
lets.rangesize(x, species_name = x@data[, 1], coordinates = "geographic",
  units = "cells")
}
\arguments{
\item{x}{a PresenceAbsence object or an SpatialPolygonsDataFrame.}

\item{species_name}{species name in the same other as in the SpatialPolygonsDataFrame (only needed if x is a SpatialPolygonsDataFrame).}

\item{coordinates}{"geographical" or "planar". Indicate wheter the shapefile has geographical or planar coordinates(only needed if x is a SpatialPolygonsDataFrame).}

\item{units}{"cell" or "squaremeters". Indicate if the units wanted are in number of cells occupied or in square meters(only needed if x is a PresenceAbsence object).}
}
\value{
The result is a matrix with the range size of each species.
If the range size accounts for the earth curvature (Yes or No) or it's unit
may differ for each argument combination:
1) SpatialPolygonsDataFrame & geographical = Square meters. Yes.
2) SpatialPolygonsDataFrame & planar = same units as the coordinates. No.
3) PresenceAbsence & cell = number of cells. No.
4) PresenceAbsence & squaremeters = squaremeters. Yes.
}
\description{
This function calculate species range size from a PresenceAbsence object or direct from the species shapefiles.
}
\examples{
\dontrun{
# SpatialPolygonsDataFrame & geographical
data(Phyllomedusa)
rangesize <- lets.rangesize(x=Phyllomedusa, coordinates="geographic")

# SpatialPolygonsDataFrame & planar
rangesize2 <- lets.rangesize(x=Phyllomedusa, coordinates="planar")

# PresenceAbsence & cell
data(PAM)
rangesize3 <- lets.rangesize(x=PAM, units="cell")

# PresenceAbsence & squaremeter
rangesize4 <- lets.rangesize(x=PAM, units="squaremeter")
}
}
\author{
Bruno Vilela
}

