\name{lets.transf}
\alias{lets.transf}
\title{Transform values of a vector}
\usage{
lets.transf(x, y, z, NUMERIC=TRUE)
}
\arguments{
  \item{x}{A vector to be transformed.}

  \item{y}{levels to be transformed.}

  \item{z}{The value to be atributed to each level (same
  order as y).}

  \item{NUMERIC}{logical, if TRUE z will be considered
  numbers.}
}
\value{
Return a vector with changed values.
}
\description{
Transform each element of a vector.
}
\author{
Bruno Vilela
}

