% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{PAM}
\alias{PAM}
\title{PresenceAbsence object for frogs of the genus Phyllomedusa}
\format{
A PresenceAbsence object
}
\source{
Generated from IUCN Spatial Data - \url{https://www.iucnredlist.org/}. 2014.
}
\usage{
PAM
}
\description{
PresenceAbsence object obtained using the function \code{\link{lets.presab}} 
for the Geographic distribution of the South American frog genus \link{Phyllomedusa}.
}
\keyword{datasets}
