% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{PAM}
\alias{PAM}
\title{PresenceAbsence object for frogs of Phyllomedusa genus}
\source{
IUCN - \url{http://www.iucnredlist.org/}. 2014.
}
\usage{
data(PAM)
}
\description{
PresenceAbsence object obtained using the function \code{\link{lets.presab}}
in the Geographic distribution of the Southern American frog genus \link{Phyllomedusa}.
}

