% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{Phyllomedusa}
\alias{Phyllomedusa}
\title{Geographic distribution of Phyllomedusa genus}
\source{
IUCN - \url{https://www.iucnredlist.org/}. 2014.
}
\usage{
data(Phyllomedusa)
}
\description{
Geographic distribution of the Southern American frog genera Phyllomedusa. 
Data was modfied from IUCN (\url{https://www.iucnredlist.org/}, downloaded in 05/2014). 
There are 32 species and 46 polygons.
}
