% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{Phyllomedusa}
\alias{Phyllomedusa}
\title{Phyllomedusa}
\format{
A simple feature collection for 32 species with 46 features and 4 fields.

\describe{
 \item{binomial}{Scientific name}
 \item{presence}{IUCN Red List distributional code}
 \item{origin}{IUCN Red List distributional code}
 \item{seasonal}{IUCN Red List distributional code}
}
}
\source{
IUCN - \url{https://www.iucnredlist.org/}. 2014.
}
\usage{
Phyllomedusa
}
\description{
Geographic distribution of the South American frog genus Phyllomedusa.
}
\details{
See the IUCN Red List Attribute Information for Species Ranges.
}
\keyword{datasets}
