\name{plot.PresenceAbsence}
\alias{plot.PresenceAbsence}
\title{Plot an object of class PresenceAbsence}
\usage{
plot(x, name=NULL, world=T)
}
\arguments{
  \item{x}{an object of class PresenceAbsence (see function
  presab).}

  \item{name}{you can specify a species to be ploted
  instead of the species richness map.}

  \item{world}{if True a map of political divisions
  (countries) is added to the plot.}
}
\description{
Plots species richness map from an object of class
PresenceAbsence or an specific species map.
}
\author{
Bruno Vilela
}
\seealso{
lets.presab

lets.presab.birds
}

