% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_PresenceAbsence.R
\name{plot.PresenceAbsence}
\alias{plot.PresenceAbsence}
\title{Plot an object of class PresenceAbsence}
\usage{
\method{plot}{PresenceAbsence}(x, name = NULL, world = TRUE, col = NULL, \dots)
}
\arguments{
\item{x}{An object of class PresenceAbsence (see function presab).}

\item{name}{A character specifying a species to be ploted instead of the complete species richness map.}

\item{world}{If \code{TURE} a map of political divisions (countries) is added to the plot.}

\item{col}{Color function (e.g. \code{\link{rainbow}}, \code{\link{heat.colors}}, \code{\link{colorRampPalette}}) to be used in the richness map.}

\item{...}{Other parameters pass to the plot function.}
}
\description{
Plots species richness map from an object of class PresenceAbsence or a
particular species' map.
}
\examples{
\dontrun{
data(PAM)
plot(PAM)
plot(PAM, xlab = "Longitude", ylab = "Latitude",
     main = "Phyllomedusa species richness")
plot(PAM, name = "Phyllomedusa atelopoides")
plot(PAM, name = "Phyllomedusa azurea")
}
}
\author{
Bruno Vilela
}
\seealso{
\code{\link{lets.presab}}

\code{\link{lets.presab.birds}}
}

