% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PresenceAbsence.R
\name{print.PresenceAbsence}
\alias{print.PresenceAbsence}
\title{Print for object of class PresenceAbsence}
\usage{
\method{print}{PresenceAbsence}(x, \dots)
}
\arguments{
\item{x}{an object of class \code{\link{PresenceAbsence}}.}

\item{...}{Other print parameters.}
}
\description{
Print for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}
