% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_PresenceAbsence.R
\name{print.summary.PresenceAbsence}
\alias{print.summary.PresenceAbsence}
\title{Print summary for object of class PresenceAbsence}
\usage{
\method{print.summary}{PresenceAbsence}(x, \dots)
}
\arguments{
\item{x}{an object of class \code{\link{PresenceAbsence}}.}

\item{...}{Other print parameters.}
}
\description{
Print summary for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}

