% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PresenceAbsence.R
\name{summary.PresenceAbsence}
\alias{summary.PresenceAbsence}
\title{Summary for object of class PresenceAbsence}
\usage{
\method{summary}{PresenceAbsence}(object)
}
\arguments{
\item{object}{an object of class \code{\link{PresenceAbsence}}.}
}
\description{
Summary for objects of class PresenceAbsence.
}
\author{
Bruno Vilela
}
