% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{temp}
\alias{temp}
\title{Avarege temperature raster for the world.}
\source{
Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. 
Very high resolution interpolated climate surfaces for global land areas. 
International Journal of Climatology 25: 1965-1978.
}
\usage{
data(temp)
}
\description{
Average temperature raster in Celsius degrees(multiplied by 100) 
for the world in 10 arc min of resolution. Data was modfied from 
WorldClim (\url{https://worldclim.com/}, downloaded in 05/2014).
}
