% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{wrld_simpl}
\alias{wrld_simpl}
\title{Simplified world country polygons}
\format{
A simple feature collection with 246 features and 11 fields.
}
\source{
Originally
 \url{https://mappinghacks.com/data/TM_WORLD_BORDERS_SIMPL-0.2.zip}, 
 now available from
 \url{https://github.com/nasa/World-Wind-Java/tree/master/WorldWind/testData/shapefiles}.
}
\usage{
wrld_simpl
}
\description{
World map in sf format. Obtained from maptools 
and converted to sf.
}
\keyword{datasets}
