% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_helps.R
\docType{data}
\name{wrld_simpl}
\alias{wrld_simpl}
\title{Simplified world country polygons}
\source{
Originally
 “http://mappinghacks.com/data/TM_WORLD_BORDERS_SIMPL-0.2.zip”, 
 now available from
 https://github.com/nasa/World-Wind-Java/tree/master/WorldWind/testData/shapefiles
}
\usage{
data(temp)
}
\description{
Shapefile in sf format. The object was obtained from maptools
 package and converted to sf.
}
